refactor(annotator): splitting canvas-related stuff into a new Workbench component
