+++ src/css/annotator.css
@@ -6,6 +6,8 @@
 
 .workbench {
   position: relative;
+  height: 600px;
+  width: 100%;
   background-color: #eeeeee;
 }
